% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_pkg.R
\name{create_pkg}
\alias{create_pkg}
\title{Create R Package}
\usage{
create_pkg(path, pkgdevt_script = TRUE, open = TRUE, ...)
}
\arguments{
\item{path}{Path to create package}

\item{pkgdevt_script}{Boolean - Should a \code{pkgdevt.R} script be created?}

\item{open}{Should the new package be opened?}

\item{...}{Passed onto \code{\link[usethis:create_package]{usethis::create_package()}}}
}
\value{
invisibly returns 0; used for side-effects.
}
\description{
This function is a wrapper around \code{\link[usethis:create_package]{usethis::create_package()}} with some
added functionality listed in details below.
}
