% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-msg_helpers.R
\name{feedback}
\alias{feedback}
\alias{msg_field}
\alias{msg_value}
\alias{msg_done}
\alias{msg_bullet}
\alias{msg_err}
\alias{msg_path}
\alias{msg_info}
\alias{msg_code}
\alias{msg_feedback}
\title{Feedback Message Utility Functions - \code{msg_}}
\usage{
msg_field(x)

msg_value(x)

msg_done(x)

msg_bullet(x, bullet = cli::symbol$bullet)

msg_err(x)

msg_path(x)

msg_info(x)

msg_code(x)

msg_feedback(x)
}
\arguments{
\item{x}{The string passed to various \code{msg_} functions.}

\item{bullet}{What to use for the message's \code{bullet}. Defaults to \code{cli::symbol$bullet}}
}
\description{
A set of helper functions for providing verbose feedback to the developer using
functions prefixed with \code{msg_} (i.e. \code{msg_error}, \code{msg_warning}, \code{msg_success}).
}
\details{
See \code{\link[=use_feedback_helpers]{use_feedback_helpers()}} to add all of these functions to your own package!
}
\seealso{
\itemize{
\item \code{\link[usethis:ui-questions]{usethis::ui-questions()}}
\item \code{\link[cli:symbol]{cli::list_symbols()}}
}

Other Feedback Utilities: 
\code{\link{indent}()},
\code{\link{inform}()},
\code{\link{use_feedback_helpers}()}
}
\concept{Feedback Utilities}
\keyword{feedback}
