% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gha_pkgdown_branch.R
\encoding{UTF-8}
\name{gha_pkgdown_branch}
\alias{gha_pkgdown_branch}
\title{Create a GitHub Actions workflow that builds a \CRANpkg{pkgdown} site}
\source{
Examples from
\href{https://github.com/r-lib/actions/tree/master/examples}{r-lib/actions}.
}
\usage{
gha_pkgdown_branch(
  pkg = ".",
  overwrite = TRUE,
  platform = "macOS",
  version = "latest"
)
}
\arguments{
\item{pkg}{Path to the package root directory.}

\item{overwrite}{Whether to overwrite an existing workflow file.}

\item{platform}{Runner operating system to use for deploying the site.
See \strong{Details}.}

\item{version}{Runner image version. See \strong{Details}.}
}
\value{
\code{\link[base:NULL]{NULL}}, invisibly, after writing a GitHub Actions
workflow to \verb{<pkg>/.github/workflows}.
}
\description{
The GitHub Actions workflow deploys a \CRANpkg{pkgdown} site for your
package to the \code{gh-pages} branch.
}
\details{
Check \url{https://github.com/actions/runner-images} to see the available
options.
}
\examples{
\dontrun{
# With Ubuntu 20.04.
gha_pkgdown_branch(platform = "ubuntu", version = "20.04")
}
}
\seealso{
\code{\link[pkgdown:build_site]{pkgdown::build_site()}} builds the package website locally.

GitHub Actions workflow helpers:
\code{\link[=gha_check_full]{gha_check_full()}},
\code{\link[=gha_update_docs]{gha_update_docs()}}
}
\concept{actions}
