% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gha_update_docs.R
\encoding{UTF-8}
\name{gha_update_docs}
\alias{gha_update_docs}
\title{Create a GitHub Actions workflow that documents and checks your package}
\source{
Examples from
\href{https://github.com/r-lib/actions/tree/master/examples}{r-lib/actions}.
}
\usage{
gha_update_docs(
  pkg = ".",
  overwrite = TRUE,
  platform = "macOS",
  version = "latest"
)
}
\arguments{
\item{pkg}{Path to the package root directory.}

\item{overwrite}{Whether to overwrite an existing workflow file.}

\item{platform}{Runner operating system to use for deploying the site.
See \strong{Details}.}

\item{version}{Runner image version. See \strong{Details}.}
}
\value{
\code{\link[base:NULL]{NULL}}, invisibly, after writing a GitHub Actions
workflow to \verb{<pkg>/.github/workflows}.
}
\description{
The GitHub Actions workflow documents your package (see \code{\link[=update_docs]{update_docs()}}),
checks it and deploys its \CRANpkg{pkgdown} site to the \code{gh-pages} branch.
}
\details{
Check \url{https://github.com/actions/runner-images} to see the available
options.
}
\examples{
\dontrun{
# With Ubuntu 20.04.
gha_update_docs(platform = "ubuntu", version = "20.04")
}
}
\seealso{
\code{\link[=update_docs]{update_docs()}} documents and checks your package.

GitHub Actions workflow helpers:
\code{\link[=gha_check_full]{gha_check_full()}},
\code{\link[=gha_pkgdown_branch]{gha_pkgdown_branch()}}
}
\concept{actions}
