% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/library_dev_packages.R
\name{library_dev_packages}
\alias{library_dev_packages}
\title{Library Package Development Packages}
\usage{
library_dev_packages(...)
}
\arguments{
\item{...}{For future use.}
}
\value{
invisible
}
\description{
This function simply libraries package development R packages into your
environment. See details for which packages are loaded in.
}
\details{
The following packages get library'd:
\itemize{
\item usethis
\item devtools
\item pkgload
\item pkgdown
\item testthat
\item pkgdown
\item testthat
\item roxygen2
\item knitr
\item pak
\item purrr
\item desc
\item chameleon
\item attachment
}
}
