% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimize_imgs.R
\name{optimize_imgs}
\alias{optimize_imgs}
\alias{optimize_imgs_all}
\title{Optimize images with reSmush.it API}
\source{
\url{https://resmush.it/features}
}
\usage{
optimize_imgs(path, outfile = path, qlty = 92)

optimize_imgs_all(
  dir = c("man/figures", "vignettes"),
  ext = "png$|jpg$",
  qlty = 92
)
}
\arguments{
\item{path}{Path to the files to optimize.}

\item{outfile}{Output file. By default is \code{path}, i.e. overwrites the file
provided.}

\item{qlty}{Integer between 0 and 100. Optimization level (JPG compression
only).}

\item{dir}{Directory or list of directories to optimize.}

\item{ext}{Regex indicating the extension of the image files to be optimized.}
}
\value{
Writes optimized image files on disk.
}
\description{
Optimize images using the reSmush.it API \url{https://resmush.it/}.
reSmush.it is a FREE API that provides image optimization. reSmush.it limits
its use to PNG, JPG, GIF, BMP and TIF pictures up to 5MB.
}
\examples{
\donttest{

# A jpg file
myjpg <- system.file("img/example.jpg", package = "pkgdev")
destjpg <- tempfile(fileext = ".jpg")

# Regular use
optimize_imgs(myjpg, destjpg)

# Use qlty
optimize_imgs(myjpg, destjpg, qlty = 20)

# With png
mypng <- system.file("img/example.png", package = "pkgdev")
destpng <- tempfile(fileext = ".png")
optimize_imgs(mypng, destpng)
}

}
