% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgdev-package.R
\name{options}
\alias{options}
\title{pkgdev Options}
\description{
Internally used, package-specific options. All options will prioritize R options() values, and fall back to environment variables if undefined. If neither the option nor the environment variable is set, a default value is used.
}
\section{Checking Option Values}{

Option values specific to \code{pkgdev} can be
accessed by passing the package name to \code{env}.

\if{html}{\out{<div class="sourceCode">}}\preformatted{options::opts(env = "pkgdev")

options::opt(x, default, env = "pkgdev")
}\if{html}{\out{</div>}}
}

\section{Options}{

\describe{
\item{pkg.author}{\describe{
Default Package Author\item{default: }{\preformatted{whoami::whoami()["fullname"]}}
\item{option: }{globalenv.pkg.author}
\item{envvar: }{R_GLOBALENV_PKG_AUTHOR (evaluated if possible, raw string otherwise)}
}}

}
}

\seealso{
options getOption Sys.getenv Sys.getenv
}
