% Generated by roxygen2 (4.0.2): do not edit by hand
\name{pkgdev.check}
\alias{pkgdev.check}
\title{Invoke R CMD check on a flat package (maybe with optional arguments, like --as-cran)}
\usage{
pkgdev.check(path, opts)
}
\arguments{
\item{opts}{Vector of strings arguments to pass to R CMD CHECK}
}
\description{
Invoke R CMD check on a flat package (maybe with optional arguments, like --as-cran)
}

