% Generated by roxygen2 (4.0.2): do not edit by hand
\name{pkgdev.clean}
\alias{pkgdev.clean}
\title{Wipe a specific package under R_HOME_USER/pkgdev/pkgs/
NOTE: when this package will be loaded again, it will be completely rebuilt}
\usage{
pkgdev.clean(pkgName)
}
\arguments{
\item{pkgName}{Name of the package to be removed}
}
\description{
Wipe a specific package under R_HOME_USER/pkgdev/pkgs/
NOTE: when this package will be loaded again, it will be completely rebuilt
}

