% Generated by roxygen2 (4.0.2): do not edit by hand
\name{pkgdev.load}
\alias{pkgdev.load}
\title{Load a package containing arbitrary file structures under R/ and src/{adapters,sources}}
\usage{
pkgdev.load(path, cc = "gcc -std=gnu99")
}
\arguments{
\item{path}{Location of the package to load}

\item{cc}{Compilator to be used (e.g. 'gcc -std=gnu99' [default])}
}
\description{
Load a package containing arbitrary file structures under R/ and src/{adapters,sources}
}

