% Generated by roxygen2 (4.0.2): do not edit by hand
\name{pkgdev.rtest}
\alias{pkgdev.rtest}
\title{Launch R unit tests (arbitrary file structure under R/tests), or display the list of test functions}
\usage{
pkgdev.rtest(path, prefix = "", show = FALSE, cc = "gcc -std=gnu99")
}
\arguments{
\item{path}{Location of the package containing tests (under /R/tests)}

\item{prefix}{Prefix for names of the functions to be tested; leave empty to test all (default)}

\item{show}{Logical, TRUE to display the list of unit tests (default: FALSE)}
}
\description{
Launch R unit tests (arbitrary file structure under R/tests), or display the list of test functions
}

