% Generated by roxygen2 (4.0.2): do not edit by hand
\name{pkgdev.tocran}
\alias{pkgdev.tocran}
\title{"Flatten" a package: gather all sources under R/ and src/ without hierarchical file structure}
\usage{
pkgdev.tocran(inPath, outPath = NULL)
}
\arguments{
\item{inPath}{Input path: location of the package to flatten}

\item{outPath}{Output path: location of the package to create [default: inPath_cran]}
}
\description{
"Flatten" a package: gather all sources under R/ and src/ without hierarchical file structure
}

