% Generated by roxygen2 (4.0.2): do not edit by hand
\name{pkgdev.unload}
\alias{pkgdev.unload}
\title{Unload a package containing arbitrary file structures under R/ and src/{adapters,sources}}
\usage{
pkgdev.unload(path)
}
\arguments{
\item{path}{Location or name of the package to unload}
}
\description{
Unload a package containing arbitrary file structures under R/ and src/{adapters,sources}
}

