% Generated by roxygen2 (4.0.2): do not edit by hand
\name{pkgdev.wipeAll}
\alias{pkgdev.wipeAll}
\title{Reset pkgdev folder under R_HOME_USER
WARNING: all loaded packages will have to be rebuilt}
\usage{
pkgdev.wipeAll()
}
\description{
Reset pkgdev folder under R_HOME_USER
WARNING: all loaded packages will have to be rebuilt
}

