% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgdown_gh_pages.R
\name{pkgdown_gh_pages}
\alias{pkgdown_gh_pages}
\title{Create a GitHub action that builds a pkgdown site}
\source{
\href{https://github.com/r-lib/actions/tree/master/examples}{r-lib/actions}
}
\usage{
pkgdown_gh_pages(
  pkg = ".",
  overwrite = TRUE,
  platform = "macOS",
  version = "latest"
)
}
\arguments{
\item{pkg}{Path to a (subdirectory of an) R package.}

\item{overwrite}{Overwrite the action if it was already present.}

\item{platform}{Platform to use for deploying the package. See Details}

\item{version}{Version of the platform. See Details.}
}
\value{
invisible
}
\description{
The GitHub action created would deploy a pkgdown site of your package
on the gh-pages branch.
}
\details{
Check \url{https://github.com/actions/virtual-environments} to see the available
options.
}
\examples{
\dontrun{
# With Ubuntu 20.04
pkgdown_gh_pages(platform = "ubuntu", version = "20.04")
}

}
\seealso{
\code{\link[=check_full_gh_action]{check_full_gh_action()}}
}
