% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/precompute_vignette.R
\encoding{UTF-8}
\name{precompute_vignette}
\alias{precompute_vignette}
\alias{precompute_vignette_all}
\title{Precompute vignettes}
\source{
Based on
\url{https://ropensci.org/blog/2019/12/08/precompute-vignettes/}.
}
\usage{
precompute_vignette(
  source,
  pkg = ".",
  figure_ext = ".png",
  create_r_file = FALSE
)

precompute_vignette_all(dir = "vignettes", pkg = ".", ...)
}
\arguments{
\item{source}{Name of the \code{.Rmd.orig} or \code{.qmd.orig} file, without the path
(e.g. \code{"some_name.Rmd.orig"} or \code{"some_name.qmd.orig"}).}

\item{pkg}{Path to the package root directory.}

\item{figure_ext}{File extension for figures plotted in the vignette.
See \strong{Details}.}

\item{create_r_file}{Whether to create an additional \R script with the code
from the vignette.}

\item{dir}{Path to the directory where the \code{.Rmd.orig} and \code{.qmd.orig} files
are stored.}

\item{...}{Additional arguments passed to \code{\link[=precompute_vignette]{precompute_vignette()}}.}
}
\value{
\code{\link[base:NULL]{NULL}}, invisibly, after precomputing the vignettes.
}
\description{
Precompute vignettes following the CRAN approach described at
\url{https://ropensci.org/blog/2019/12/08/precompute-vignettes/}.
}
\details{
This function reads vignette source files from the package's \code{vignettes}
directory and moves plots from the package root to that directory.

\code{\link[=precompute_vignette]{precompute_vignette()}} processes the files named in \code{source}.
\code{\link[=precompute_vignette_all]{precompute_vignette_all()}} finds and processes every \code{.Rmd.orig} and
\code{.qmd.orig} file in \code{dir}.
\subsection{Important}{

In your \code{.Rmd.orig} or \code{.qmd.orig} file, set the following chunk option when
producing plots:

\if{html}{\out{<div class="sourceCode r">}}\preformatted{knitr::opts_chunk$set(
  ...,
  fig.path = "./",
  ...,
)

}\if{html}{\out{</div>}}
}
}
\examples{
\dontrun{
precompute_vignette(source = "precompute.Rmd.orig")
}
}
\seealso{
\code{\link[=update_docs]{update_docs()}} runs the broader package maintenance workflow.

Documentation rendering helpers:
\code{\link[=build_qmd]{build_qmd()}}
}
\concept{renderers}
