% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/precompute_vignette.R
\name{precompute_vignette}
\alias{precompute_vignette}
\title{Precompute vignettes}
\source{
\url{https://ropensci.org/blog/2019/12/08/precompute-vignettes/}
}
\usage{
precompute_vignette(source, figure_ext = ".png")
}
\arguments{
\item{source}{Name without path of the \code{.Rmd.orig} file
(e.g. \code{"some_name.Rmd.orig"}).}

\item{figure_ext}{Extension of the figures plotted on the vignette.
See Details.}
}
\value{
A precomputed vignette
}
\description{
Precompute vignettes from CRAN, based on
\url{https://ropensci.org/blog/2019/12/08/precompute-vignettes/}.
}
\details{
The function would search for the desired precomputed vignette on the
\code{"./vignettes/"} folder and for plots on the root \code{"./"} folder.
\subsection{Important}{

On your \code{.Rmd.orig} file make sure you have set at least the following
lines if you are producing plots:\if{html}{\out{<div class="r">}}\preformatted{knitr::opts_chunk$set(
  ... ,
  fig.path = "./",
  ... ,
  )

}\if{html}{\out{</div>}}
}
}
\examples{
\dontrun{

precompute_vignette(source = "precompute.Rmd.orig")
}

}
