% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-msg_helpers.R
\name{use_feedback_helpers}
\alias{use_feedback_helpers}
\title{Use Feedback Helpers}
\usage{
use_feedback_helpers(pkg = ".", open = TRUE)
}
\arguments{
\item{pkg}{Package - defaults to "."}
}
\value{
invisibly returns nothing - used for side-effects.
}
\description{
This function allows a package developer to automatically add the suite of
feedback \code{msg_} functions utilized in this package to their own package.
}
\details{
Note that this adds a new file to your package under \code{R/utils-msg_helpers.R}
as well as runs \code{devtools::document()} to document the added functions and
their corresponding dependencies.

The functions added are:
\itemize{
\item \code{\link[=msg_field]{msg_field()}}
\item \code{\link[=msg_value]{msg_value()}}
\item \code{\link[=msg_path]{msg_path()}}
\item \code{\link[=msg_done]{msg_done()}}
\item \code{\link[=msg_bullet]{msg_bullet()}}
\item \code{\link[=msg_code]{msg_code()}}
\item \code{\link[=msg_feedback]{msg_feedback()}}
\item \code{\link[=msg_err]{msg_err()}}
\item \code{\link[=msg_info]{msg_info()}}
}

plus, \code{\link[=inform]{inform()}} and \code{\link[=indent]{indent()}}.
}
\seealso{
\itemize{
\item \code{\link[usethis:ui-questions]{usethis::ui-questions()}}
\item \code{\link[cli:symbol]{cli::list_symbols()}}
}

Other Feedback Utilities: 
\code{\link{feedback}},
\code{\link{indent}()},
\code{\link{inform}()}
}
\concept{Feedback Utilities}
\keyword{feedback}
