% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_global_gitgnore.R
\encoding{UTF-8}
\name{add_global_gitgnore}
\alias{add_global_gitgnore}
\title{Add a global \code{.gitignore} file to a package}
\usage{
add_global_gitgnore(pkg = ".")
}
\arguments{
\item{pkg}{Path to a (subdirectory of an) R package.}
}
\value{
Invisibly returns \code{NULL} after writing a global \code{.gitignore} file.
}
\description{
Adds a global \code{.gitignore} file to the package. The file is based on the
default \code{.gitignore} provided by GitHub.
}
\examples{
\dontrun{
add_global_gitgnore()
}
}
\seealso{
\itemize{
\item \code{\link[usethis:use_git_ignore]{usethis::use_git_ignore()}} adds entries to \code{.gitignore}.
\item \code{\link[usethis:git_vaccinate]{usethis::git_vaccinate()}} adds global Git ignore rules.
}

Package maintenance helpers:
\code{\link[=update_docs]{update_docs()}}
}
\concept{maintenance}
