% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_global_gitgnore.R
\name{add_global_gitgnore}
\alias{add_global_gitgnore}
\title{Add \code{.gitignore} to package.}
\usage{
add_global_gitgnore(pkg = ".")
}
\arguments{
\item{pkg}{Path to a (subdirectory of an) R package.}
}
\value{
Invisible, and writes a global \code{.gitignore} file
}
\description{
Adds a global \code{.gitgnore} file to the package. File based on default
\code{.gitignore} provided by GitHub.
}
\examples{
\dontrun{
add_global_gitgnore()
}

}
\seealso{
\code{\link[usethis:use_git_ignore]{usethis::use_git_ignore()}}, \code{\link[usethis:git_vaccinate]{usethis::git_vaccinate()}}
}
