% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_global_gitgnore.R
\name{add_global_gitgnore}
\alias{add_global_gitgnore}
\title{Add .gitignore to package}
\usage{
add_global_gitgnore(pkg = ".")
}
\arguments{
\item{pkg}{Path to a (subdirectory of an) R package.}
}
\value{
A .gitignore file
}
\description{
Adds a global .gitgnore file to the package. File based on default
.gitignore provided by GitHub.
}
\examples{
\dontrun{
add_global_gitgnore()
}

}
\seealso{
\code{\link[usethis:use_git_ignore]{usethis::use_git_ignore()}}, \code{\link[usethis:git_vaccinate]{usethis::git_vaccinate()}}
}
