% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_qmd.R
\encoding{UTF-8}
\name{build_qmd}
\alias{build_qmd}
\alias{build_readme_qmd}
\title{Build Quarto files for a package}
\usage{
build_qmd(files, path = ".", ..., quiet = TRUE)

build_readme_qmd(path = ".", quiet = TRUE, ...)
}
\arguments{
\item{files}{Quarto files to be rendered.}

\item{path}{Path to the top-level directory of the source package.}

\item{...}{Additional arguments passed to \code{\link[quarto:quarto_render]{quarto::quarto_render()}}.}

\item{quiet}{If \code{TRUE}, suppresses most output. Set to \code{FALSE}
if you need to debug.}
}
\value{
\code{TRUE}, invisibly.
}
\description{
\code{build_qmd()} is a wrapper around \code{\link[quarto:quarto_render]{quarto::quarto_render()}} that first
installs a temporary copy of the package, then renders each Quarto file in a
clean \R session. \code{build_readme_qmd()} locates your \code{README.qmd} and builds
it into a \code{README.md}.
}
\seealso{
\itemize{
\item \code{\link[=build_readme_qmd]{build_readme_qmd()}} builds \code{README.qmd} files.
\item \code{\link[=precompute_vignette]{precompute_vignette()}} precomputes vignettes.
\item \code{\link[devtools:build_readme]{devtools::build_readme()}} builds \code{README} files from R Markdown.
}

Documentation renderers:
\code{\link[=precompute_vignette]{precompute_vignette()}}
}
\concept{renderers}
