% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_qmd.R
\name{build_qmd}
\alias{build_qmd}
\alias{build_readme_qmd}
\title{Build a Quarto files package}
\usage{
build_qmd(files, path = ".", ..., quiet = TRUE)

build_readme_qmd(path = ".", quiet = TRUE, ...)
}
\arguments{
\item{files}{The Quarto files to be rendered.}

\item{path}{path to the package to build the readme.}

\item{...}{additional arguments passed to \code{\link[quarto:quarto_render]{quarto::quarto_render()}}.}

\item{quiet}{If \code{TRUE}, suppress output.}
}
\description{
\code{build_qmd()} is a wrapper around \code{\link[quarto:quarto_render]{quarto::quarto_render()}} that first
installs a temporary copy of the package, and then renders each .Rmd in a
clean R session. \code{build_readme_qmd()} locates your README.and builds it
into a README.md
}
\seealso{
\code{\link[devtools:build_rmd]{devtools::build_readme()}}
}
\concept{functions for rendering documentation}
