% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_rd_titles.R
\name{check_rd_titles}
\alias{check_rd_titles}
\title{Check the titles of a Rd file}
\usage{
check_rd_titles(pkg = ".")
}
\arguments{
\item{pkg}{Path to a (subdirectory of an) R package.}
}
\value{
A data frame with the results.
}
\description{
Check if title of Rd files ends in a period.
}
\examples{
\dontrun{

check_rd_titles()
}

}
