% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gha_check_full.R
\encoding{UTF-8}
\name{gha_check_full}
\alias{gha_check_full}
\title{Create a GitHub Actions workflow that checks your package regularly}
\source{
Examples from
\href{https://github.com/r-lib/actions/tree/master/examples}{r-lib/actions}.
}
\usage{
gha_check_full(pkg = ".", overwrite = TRUE, cron_expr = "30 08 1 * *")
}
\arguments{
\item{pkg}{Path to a (subdirectory of an) R package.}

\item{overwrite}{Whether to overwrite an existing action.}

\item{cron_expr}{A valid cron expression. Defaults to 08:30 AM on the first
day of the month. See \strong{Details}.}
}
\value{
Invisibly returns \code{NULL} after writing a GitHub Actions workflow to
\verb{<pkg>/.github/workflows}.
}
\description{
The GitHub Actions workflow runs \verb{R CMD check} on your package.
It uses a wide range of platforms, which can be reduced by commenting out or
deleting platforms in the matrix configuration.
}
\details{
Use \href{https://crontab.guru/#30_08_1_*_*}{crontab.guru} to check and
create your own cron expression.
}
\examples{
\dontrun{
gha_check_full(cron_expr = "57 16 12 * *")
}
}
\seealso{
\code{\link[usethis:use_github_action]{usethis::use_github_action()}} creates GitHub Actions workflows.

GitHub Actions workflow helpers:
\code{\link[=gha_pkgdown_branch]{gha_pkgdown_branch()}},
\code{\link[=gha_update_docs]{gha_update_docs()}}
}
\concept{actions}
