% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gha_check_full.R
\encoding{UTF-8}
\name{gha_check_full}
\alias{gha_check_full}
\title{Create a GitHub Actions workflow that checks your package regularly}
\source{
Examples from
\href{https://github.com/r-lib/actions/tree/master/examples}{r-lib/actions}.
}
\usage{
gha_check_full(pkg = ".", overwrite = TRUE, cron_expr = "30 08 1 * *")
}
\arguments{
\item{pkg}{Path to a (subdirectory of an) R package.}

\item{overwrite}{Overwrite the action if it was already present.}

\item{cron_expr}{Valid cron expression. By default, the first day of the
month at 08:30 AM. See \strong{Details}.}
}
\value{
Invisibly returns \code{NULL} after writing a GitHub Actions workflow to
\verb{<pkg>/.github/workflows}.
}
\description{
The GitHub Actions workflow runs \verb{R CMD check} on your package.
It uses a wide range of platforms, which can be reduced by commenting out or
deleting platforms in the matrix configuration.
}
\details{
Use \href{https://crontab.guru/#30_08_1_*_*}{crontab.guru} to check and
create your own cron tag.
}
\examples{
\dontrun{
gha_check_full(cron_expr = "57 16 12 * *")
}
}
\seealso{
\itemize{
\item \code{\link[=gha_pkgdown_branch]{gha_pkgdown_branch()}} creates a \CRANpkg{pkgdown} deployment action.
\item \code{\link[=gha_update_docs]{gha_update_docs()}} creates a documentation and deployment action.
\item \code{\link[usethis:use_github_action]{usethis::use_github_action()}} creates GitHub Actions workflows.
}

GitHub Actions helpers:
\code{\link[=gha_pkgdown_branch]{gha_pkgdown_branch()}},
\code{\link[=gha_update_docs]{gha_update_docs()}}
}
\concept{actions}
