% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gha_check_full.R
\name{gha_check_full}
\alias{gha_check_full}
\title{Create a GitHub action that checks regularly your package}
\source{
\href{https://github.com/r-lib/actions/tree/master/examples}{r-lib/actions}
}
\usage{
gha_check_full(pkg = ".", overwrite = TRUE, cron_expr = "30 08 1 * *")
}
\arguments{
\item{pkg}{Path to a (subdirectory of an) R package.}

\item{overwrite}{Overwrite the action if it was already present.}

\item{cron_expr}{Valid cron expression. By default, the first
day of the month at 08:30 AM. See Details.}
}
\value{
A GitHub Action on \verb{<pkg>/.github/workflows}.
}
\description{
The GitHub action created would run \verb{R CMD check} on your package.
It uses a wide range of platforms, can be reduced by commenting or deleting
platforms on the matrix config.
}
\details{
Use \href{https://crontab.guru/#30_08_1_*_*}{crontab.guru} to check and
create your own cron tag.
}
\examples{
\dontrun{
gha_check_full(cron_expr = "57 16 12 * *")
}

}
\seealso{
\code{\link[usethis:use_github_action]{usethis::use_github_action_check_full()}}
}
