% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gha_pkgdown_branch.R
\encoding{UTF-8}
\name{gha_pkgdown_branch}
\alias{gha_pkgdown_branch}
\title{Create a GitHub Actions workflow that builds a \CRANpkg{pkgdown} site}
\source{
Examples from
\href{https://github.com/r-lib/actions/tree/master/examples}{r-lib/actions}.
}
\usage{
gha_pkgdown_branch(
  pkg = ".",
  overwrite = TRUE,
  platform = "macOS",
  version = "latest"
)
}
\arguments{
\item{pkg}{Path to a (subdirectory of an) R package.}

\item{overwrite}{Whether to overwrite an existing action.}

\item{platform}{Platform to use for deploying the package. See \strong{Details}.}

\item{version}{Version of the platform. See \strong{Details}.}
}
\value{
Invisibly returns \code{NULL} after writing a GitHub Actions workflow to
\verb{<pkg>/.github/workflows}.
}
\description{
The GitHub Actions workflow deploys a \CRANpkg{pkgdown} site for your
package on the \code{gh-pages} branch.
}
\details{
Check \url{https://github.com/actions/runner-images} to see the available
options.
}
\examples{
\dontrun{
# With Ubuntu 20.04
gha_pkgdown_branch(platform = "ubuntu", version = "20.04")
}
}
\seealso{
GitHub Actions workflow helpers:
\code{\link[=gha_check_full]{gha_check_full()}},
\code{\link[=gha_update_docs]{gha_update_docs()}}
}
\concept{actions}
