% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgdev-package.R
\docType{package}
\name{pkgdev-package}
\alias{pkgdev}
\alias{pkgdev-package}
\title{pkgdev: Helpers to Develop Packages with GitHub Actions}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Helpers for package maintenance, documentation rendering and GitHub Actions workflows. The package is primarily intended for personal use, but you are welcome to use it at your own risk.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://dieghernan.github.io/pkgdev/}
  \item \url{https://github.com/dieghernan/pkgdev}
  \item Report bugs at \url{https://github.com/dieghernan/pkgdev/issues}
}

}
\author{
\strong{Maintainer}: Diego Hernangómez \email{diego.hernangomezherrero@gmail.com} (\href{https://orcid.org/0000-0001-8457-4658}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Diego Hernangómez \email{diego.hernangomezherrero@gmail.com} (\href{https://orcid.org/0000-0001-8457-4658}{ORCID}) [copyright holder]
}

}
\keyword{internal}
