% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgdev-package.R
\docType{package}
\name{pkgdev-package}
\alias{pkgdev}
\alias{pkgdev-package}
\title{pkgdev: Helpers to Develop a Package using GitHub Actions}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

A small set of functions that takes advantage of GitHub Actions for making your life easier as a R package developer. This package is primarily intended for personal use, however feel free to use it (at your own risk :)).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://dieghernan.github.io/pkgdev/}
  \item \url{https://github.com/dieghernan/pkgdev}
  \item Report bugs at \url{https://github.com/dieghernan/pkgdev/issues}
}

}
\author{
\strong{Maintainer}: Diego Hernangómez \email{diego.hernangomezherrero@gmail.com} (\href{https://orcid.org/0000-0001-8457-4658}{ORCID}) [copyright holder]

}
\keyword{internal}
