% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgdev-package.R
\docType{package}
\name{pkgdev-package}
\alias{pkgdev}
\alias{pkgdev-package}
\title{pkgdev: Suite of Helpers for the R Package Developer}
\description{
Automate and expedite your package development process in R.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jimbrig/pkgdev}
  \item \url{http://docs.jimbrig.com/pkgdev/}
  \item Report bugs at \url{https://github.com/jimbrig/pkgdev/issues}
}

}
\author{
\strong{Maintainer}: Jimmy Briggs \email{jimmy.briggs@jimbrig.com} (\href{https://orcid.org/0000-0002-7489-8787}{ORCID})

}
\keyword{internal}
