% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/precompute_vignette.R
\encoding{UTF-8}
\name{precompute_vignette}
\alias{precompute_vignette}
\alias{precompute_vignette_all}
\title{Precompute vignettes}
\source{
Based on
\url{https://ropensci.org/blog/2019/12/08/precompute-vignettes/}.
}
\usage{
precompute_vignette(
  source,
  pkg = ".",
  figure_ext = ".png",
  create_r_file = FALSE
)

precompute_vignette_all(dir = "vignettes", pkg = ".", ...)
}
\arguments{
\item{source}{Name of the \code{.Rmd.orig} or \code{.qmd.orig} file, without the path
(e.g. \code{"some_name.Rmd.orig"} or \code{"some_name.qmd.orig"}).}

\item{pkg}{Path to a (subdirectory of an) R package.}

\item{figure_ext}{Extension of the figures plotted in the vignette.
See \strong{Details}.}

\item{create_r_file}{Whether to create an additional \R script with the code
of the vignette.}

\item{dir}{Path to the directory where the \code{.Rmd.orig} and \code{.qmd.orig} files
are stored.}

\item{...}{Parameters passed to \code{\link[=precompute_vignette]{precompute_vignette()}}.}
}
\value{
Invisibly returns \code{NULL} after writing a precomputed vignette.
}
\description{
Precompute vignettes using the CRAN approach, based on
\url{https://ropensci.org/blog/2019/12/08/precompute-vignettes/}.
}
\details{
This function searches for the desired precomputed vignette in the
\code{"./vignettes/"} directory and for plots in the root \code{"./"} directory.
\subsection{Important}{

In your \code{.Rmd.orig} or \code{.qmd.orig} file, make sure you have set at least the
following lines if you are producing plots:

\if{html}{\out{<div class="sourceCode r">}}\preformatted{knitr::opts_chunk$set(
  ...,
  fig.path = "./",
  ...,
)

}\if{html}{\out{</div>}}
}
}
\examples{
\dontrun{
precompute_vignette(source = "precompute.Rmd.orig")
}
}
\seealso{
\itemize{
\item \code{\link[=build_qmd]{build_qmd()}} builds Quarto files.
\item \code{\link[=build_readme_qmd]{build_readme_qmd()}} builds \code{README.qmd} files.
}

Documentation renderers:
\code{\link[=build_qmd]{build_qmd()}}
}
\concept{renderers}
