% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/precompute_vignette.R
\name{precompute_vignette}
\alias{precompute_vignette}
\alias{precompute_vignette_all}
\title{Precompute vignettes}
\source{
\url{https://ropensci.org/blog/2019/12/08/precompute-vignettes/}
}
\usage{
precompute_vignette(source, figure_ext = ".png", create_r_file = FALSE)

precompute_vignette_all(dir = "vignettes", ...)
}
\arguments{
\item{source}{Name without path of the \code{.Rmd.orig} file
(e.g. \code{"some_name.Rmd.orig"}).}

\item{figure_ext}{Extension of the figures plotted on the vignette.
See \strong{Details}.}

\item{create_r_file}{Wheter to create an additional R file with the code
of the vignette.}

\item{dir}{Path to directory where the "Rmd.orig" files
are stored.}

\item{...}{Parameters passed to \code{\link[=precompute_vignette]{precompute_vignette()}}.}
}
\value{
A precomputed vignette
}
\description{
Precompute vignettes from CRAN, based on
\url{https://ropensci.org/blog/2019/12/08/precompute-vignettes/}.
}
\details{
The function would search for the desired precomputed vignette on the
\code{"./vignettes/"} folder and for plots on the root \code{"./"} folder.
\subsection{Important}{

On your \code{.Rmd.orig} file make sure you have set at least the following
lines if you are producing plots:

\if{html}{\out{<div class="sourceCode r">}}\preformatted{knitr::opts_chunk$set(
  ... ,
  fig.path = "./",
  ... ,
  )

}\if{html}{\out{</div>}}
}
}
\examples{
\dontrun{

precompute_vignette(source = "precompute.Rmd.orig")
}

}
