% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_docs.R
\encoding{UTF-8}
\name{update_docs}
\alias{update_docs}
\title{Document your package}
\usage{
update_docs(
  pkg = ".",
  url_update = TRUE,
  create_codemeta = TRUE,
  create_cff = TRUE,
  build_readme = TRUE,
  verbose = TRUE,
  precompute = TRUE,
  opt_imgs = TRUE,
  opt_dir = c("man/figures", "vignettes"),
  opt_ext = "png$|jpg$",
  opt_overwrite = TRUE,
  add_contributors = TRUE,
  ...
)
}
\arguments{
\item{pkg}{Path to a (subdirectory of an) R package.}

\item{url_update}{A logical value. Should URLs be updated with
\code{\link[urlchecker:url_update]{urlchecker::url_update()}}?}

\item{create_codemeta}{A logical value. Should \code{codemeta.json} be created
with \code{\link[codemetar:write_codemeta]{codemetar::write_codemeta()}}?}

\item{create_cff}{A logical value. Should \code{CITATION.cff} be created with
\code{\link[cffr:cff_write]{cffr::cff_write()}}?}

\item{build_readme}{A logical value. Should \code{README.qmd}, \code{README.Rmd} or
pkgdown index files be built?}

\item{verbose}{A logical value. Should informative messages be displayed on
the console?}

\item{precompute}{A logical value. Should vignettes be detected and
precomputed? See also \code{\link[=precompute_vignette]{precompute_vignette()}}.}

\item{opt_imgs}{A logical value. Should images be optimized with
\code{\link[resmush:resmush_dir]{resmush::resmush_dir()}}?}

\item{opt_dir, opt_ext, opt_overwrite}{Options passed to the \code{dir}, \code{ext} and
\code{overwrite} arguments of \code{\link[resmush:resmush_dir]{resmush::resmush_dir()}}.}

\item{add_contributors}{Deprecated.}

\item{...}{Additional arguments passed to downstream functions.}
}
\value{
Invisibly returns \code{NULL}. When \code{verbose = TRUE}, it emits progress
messages.
}
\description{
Run routine checks on the package:
\itemize{
\item Clean \code{DESCRIPTION} with \code{\link[usethis:use_tidy_description]{usethis::use_tidy_description()}}.
\item Compress data in the \code{"./data"} and \code{"./R"} paths using
\code{\link[tools:resaveRdaFiles]{tools::resaveRdaFiles()}}.
\item Style code with \code{\link[styler:style_pkg]{styler::style_pkg()}}.
\item Check URLs with \code{\link[urlchecker:url_check]{urlchecker::url_check()}}.
\item Roxygenize with \code{\link[roxygen2:roxygenise]{roxygen2::roxygenise()}}.
\item Precompute vignettes, if present
(see \url{https://ropensci.org/blog/2019/12/08/precompute-vignettes/}).
\item Rebuild \code{README.qmd}, \code{README.Rmd} or pkgdown index files, if present.
\item Optimize images with \code{\link[resmush:resmush_dir]{resmush::resmush_dir()}}.
\item Write \code{codemeta.json} with \code{\link[codemetar:write_codemeta]{codemetar::write_codemeta()}}.
\item Write \code{CITATION.cff} with \code{\link[cffr:cff_write]{cffr::cff_write()}}.
\item Write a project-level Codex configuration file.
}
}
\details{
This function updates and cleans the package using a mix of best practices,
such as checking URLs, roxygenizing and rebuilding the \code{README}. It also
applies discretionary practices such as the tidyverse approach for the
\code{DESCRIPTION} file, overall code style and \code{codemeta.json}.
}
\examples{
\dontrun{
update_docs()
}
}
\seealso{
\itemize{
\item \code{\link[=add_global_gitgnore]{add_global_gitgnore()}} updates package ignore files.
\item \code{\link[=build_qmd]{build_qmd()}} builds Quarto files.
\item \code{\link[=check_rd_titles]{check_rd_titles()}} checks generated Rd titles.
\item \code{\link[=precompute_vignette]{precompute_vignette()}} precomputes vignettes.
\item \code{\link[usethis:use_tidy_description]{usethis::use_tidy_description()}} cleans \code{DESCRIPTION}.
\item \code{\link[styler:style_pkg]{styler::style_pkg()}} styles package code.
\item \code{\link[urlchecker:url_check]{urlchecker::url_check()}} checks URLs.
\item \code{\link[roxygen2:roxygenise]{roxygen2::roxygenise()}} updates documentation.
\item \code{\link[devtools:build_readme]{devtools::build_readme()}} builds \code{README.Rmd} files.
\item \code{\link[codemetar:write_codemeta]{codemetar::write_codemeta()}} writes \code{codemeta.json}.
\item \code{\link[tools:resaveRdaFiles]{tools::resaveRdaFiles()}} compresses data files.
}

Package maintenance helpers:
\code{\link[=add_global_gitgnore]{add_global_gitgnore()}}
}
\concept{maintenance}
